in_syslog: fix buffer space check and other cleanups

+++ plugins/in_syslog/syslog_server.c
@@ -112,8 +112,10 @@ int syslog_server_create(struct flb_syslog *ctx)
 int syslog_server_destroy(struct flb_syslog *ctx)
 {
     if (ctx->mode == FLB_SYSLOG_UNIX_TCP || ctx->mode == FLB_SYSLOG_UNIX_UDP) {
-        unlink(ctx->unix_path);
-        flb_free(ctx->unix_path);
+        if (ctx->unix_path) {
+            unlink(ctx->unix_path);
+            flb_free(ctx->unix_path);
+        }
     }
     else {
         flb_free(ctx->listen);
